Changed doubles to floats to increase performance and added a txt file with results and notes.

+++ src/cpu/color.h
@@ -7,9 +7,9 @@
 
 using color = vec3;
 
-inline double linear_to_gamma(double linear_component){
+inline float linear_to_gamma(float linear_component){
     if (linear_component > 0)
-        return std::sqrt(linear_component);
+        return std::sqrtf(linear_component);
     return 0;
 }
